LayoutTests/media/modern-media-controls/skip-back-button/skip-back-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/skip-back-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>SkipBackButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const skipBackButton = new SkipBackButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("skipBackButton.element.localName", "button");
shouldBeTrue('skipBackButton.element.classList.contains("icon")');
shouldBeTrue('skipBackButton.element.classList.contains("skip-back")');
shouldBeEqualToString("skipBackButton.iconName", `${Icons.SkipBack}`);

const iOSSkipBackButton = new SkipBackButton({ layoutTraits: LayoutTraits.iOS });

scheduler.frameDidFire = function()
{
    if ([skipBackButton, iOSSkipBackButton].some(button => button.width == 0))
        return;

    shouldBeTrue('skipBackButton.element.style.webkitMaskImage.includes("macOS/interval-skip-back@")');
    shouldBeTrue('iOSSkipBackButton.element.style.webkitMaskImage.includes("iOS/interval-skip-back@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
